[refactoring][10] improve settings.py
- remove settings_dev (as it is redundent with ENV which already
  let us include dev specific settings)
- validate ENV setting (either PROD or DEV otherwise exception)
- removed dev setting override for `SESSION_COOKIE_HTTPONLY`
  (why did we need this override ?)
- removed dev setting override for `CORS_ALLOW_CREDENTIALS`
  (why did we need this override ?)
