+++ docker-compose.dev.yml
@@ -19,9 +19,6 @@ services:
             - "postgres"
         ports:
             - "${DJANGO_DOCKER_PORT}:9000"
-        environment:
-            # Debug settings
-            - DJANGO_SETTINGS_MODULE=geocity.settings_dev
 
     postgres:
         image: postgis/postgis:13-3.2
